+++ Flow.Launcher.Infrastructure/Win32Helper.cs
@@ -1,10 +1,13 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Globalization;
+using System.Linq;
 using System.Runtime.InteropServices;
 using System.Windows;
 using System.Windows.Interop;
+using System.Windows.Markup;
 using System.Windows.Media;
 using Flow.Launcher.Infrastructure.UserSettings;
 using Microsoft.Win32;
@@ -14,6 +17,7 @@ using Windows.Win32.Graphics.Dwm;
 using Windows.Win32.UI.Input.KeyboardAndMouse;
 using Windows.Win32.UI.WindowsAndMessaging;
 using Point = System.Windows.Point;
+using SystemFonts = System.Windows.SystemFonts;
 
 namespace Flow.Launcher.Infrastructure
 {
@@ -592,6 +596,74 @@ namespace Flow.Launcher.Infrastructure
             {
                 // Ignored
             }
+        }
+
+        #endregion
+
+        #region System Font
+
+        private static readonly Dictionary<string, string> _languageToNotoSans = new()
+        {
+            { "ko", "Noto Sans KR" },
+            { "ja", "Noto Sans JP" },
+            { "zh-CN", "Noto Sans SC" },
+            { "zh-SG", "Noto Sans SC" },
+            { "zh-Hans", "Noto Sans SC" },
+            { "zh-TW", "Noto Sans TC" },
+            { "zh-HK", "Noto Sans TC" },
+            { "zh-MO", "Noto Sans TC" },
+            { "zh-Hant", "Noto Sans TC" },
+            { "th", "Noto Sans Thai" },
+            { "ar", "Noto Sans Arabic" },
+            { "he", "Noto Sans Hebrew" },
+            { "hi", "Noto Sans Devanagari" },
+            { "bn", "Noto Sans Bengali" },
+            { "ta", "Noto Sans Tamil" },
+            { "el", "Noto Sans Greek" },
+            { "ru", "Noto Sans" },
+            { "en", "Noto Sans" },
+            { "fr", "Noto Sans" },
+            { "de", "Noto Sans" },
+            { "es", "Noto Sans" },
+            { "pt", "Noto Sans" }
+        };
+
+        public static string GetSystemDefaultFont()
+        {
+            try
+            {
+                var culture = CultureInfo.CurrentCulture;
+                var language = culture.Name; // e.g., "zh-TW"
+                var langPrefix = language.Split('-')[0]; // e.g., "zh"
+
+                // First, try to find by full name, and if not found, fallback to prefix
+                if (TryGetNotoFont(language, out var notoFont) || TryGetNotoFont(langPrefix, out notoFont))
+                {
+                    // If the font is installed, return it
+                    if (Fonts.SystemFontFamilies.Any(f => f.Source.Equals(notoFont)))
+                    {
+                        return notoFont;
+                    }
+                }
+
+                // If Noto font is not found, fallback to the system default font
+                var font = SystemFonts.MessageFontFamily;
+                if (font.FamilyNames.TryGetValue(XmlLanguage.GetLanguage("en-US"), out var englishName))
+                {
+                    return englishName;
+                }
+
+                return font.Source ?? "Segoe UI";
+            }
+            catch
+            {
+                return "Segoe UI";
+            }
+        }
+
+        private static bool TryGetNotoFont(string langKey, out string notoFont)
+        {
+            return _languageToNotoSans.TryGetValue(langKey, out notoFont);
         }
 
         #endregion
